requirements.txt
fastapi==0.111.0
starlette==0.37.2
pydantic==2.7.4
uvicorn[standard]==0.30.1
python-dotenv==1.0.1

langchain==0.2.7
langchain-core>=0.2.17,<0.3.0
langserve==0.2.2
langchain-openai==0.1.16
langchain-qdrant==0.1.2
qdrant-client>=1.10.1,<2.0.0
sse-starlette==2.4.1
httpx==0.27.2

# Nota: No incluimos dependencias de OCR/processing pesadas en la imagen del servidor
# porque el índice ya debe existir en Qdrant (se crea desde el notebook).


